Server_l1. Настроил окно авторизации. добавил переключение на окно регистрации.
