--- chat-client/src/main/resources/Authorization.fxml
@@ -1,99 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-
-<AnchorPane prefHeight="160.0" prefWidth="200.0" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ru.gb.Chatterbox.client.Authorization">
-   <VBox fx:id="loginPanel" prefHeight="200.0" prefWidth="200.0">
-      <Label text="Login:">
-         <padding>
-            <Insets top="5.0" />
-         </padding>
-         <VBox.margin>
-            <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-         </VBox.margin></Label>
-      <TextField fx:id="LoginField">
-         <VBox.margin>
-            <Insets left="5.0" right="5.0" />
-         </VBox.margin></TextField>
-      <Label text="Password:">
-         <VBox.margin>
-            <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-         </VBox.margin></Label>
-      <TextField fx:id="PasswordField" onAction="#sendAuth">
-         <VBox.margin>
-            <Insets bottom="5.0" left="5.0" right="5.0" />
-         </VBox.margin>
-      </TextField>
-      <Button onAction="#sendAuth" text="Connect">
-         <VBox.margin>
-            <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-         </VBox.margin></Button>
-      <Label accessibleRole="BUTTON" alignment="TOP_RIGHT" contentDisplay="BOTTOM" onMouseClicked="#sendRegistrationWindow" prefHeight="26.0" prefWidth="150.0" text="Registration" textAlignment="RIGHT">
-         <VBox.margin>
-            <Insets bottom="5.0" left="45.0" top="25.0" />
-         </VBox.margin>
-         <opaqueInsets>
-            <Insets />
-         </opaqueInsets>
-      </Label>
-   </VBox>
-
-
-   <VBox fx:id="registrationPanel" prefHeight="200.0" prefWidth="200.0" visible="false">
-      <children>
-         <Label text="Login:">
-            <VBox.margin>
-               <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-            </VBox.margin>
-            <padding>
-               <Insets top="5.0" />
-            </padding>
-         </Label>
-         <TextField fx:id="newLoginField">
-            <VBox.margin>
-               <Insets left="5.0" right="5.0" />
-            </VBox.margin>
-         </TextField>
-         <Label text="Password:">
-            <VBox.margin>
-               <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-            </VBox.margin>
-         </Label>
-         <TextField fx:id="newPasswordField" onAction="#sendReg">
-            <VBox.margin>
-               <Insets bottom="5.0" left="5.0" right="5.0" />
-            </VBox.margin>
-         </TextField>
-         <Label text="Nickname">
-            <VBox.margin>
-               <Insets bottom="5.0" left="5.0" right="5.0" />
-            </VBox.margin>
-         </Label>
-         <TextField fx:id="newNickField" onAction="#sendReg">
-            <VBox.margin>
-               <Insets bottom="5.0" left="5.0" right="5.0" />
-            </VBox.margin>
-         </TextField>
-         <HBox prefHeight="100.0" prefWidth="200.0">
-            <children>
-               <Button onAction="#sendReg" text="Registration">
-                  <HBox.margin>
-                     <Insets left="5.0" right="10.0" />
-                  </HBox.margin>
-               </Button>
-               <Label accessibleRole="BUTTON" alignment="BOTTOM_RIGHT" contentDisplay="BOTTOM" onMouseClicked="#sendAuthorisationWindow" prefHeight="26.0" prefWidth="100.0" text="Authorization" textAlignment="RIGHT">
-                  <HBox.margin>
-                     <Insets right="5.0" />
-                  </HBox.margin>
-               </Label>
-            </children>
-         </HBox>
-      </children>
-   </VBox>
-</AnchorPane>
