+++ chat-client/src/main/resources/ChatWindow.fxml
@@ -3,29 +3,115 @@
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.ButtonBar?>
-<?import javafx.scene.control.ListView?>
+<?import javafx.scene.control.Label?>
 <?import javafx.scene.control.Menu?>
 <?import javafx.scene.control.MenuBar?>
 <?import javafx.scene.control.MenuItem?>
 <?import javafx.scene.control.TextArea?>
 <?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.TreeView?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
 
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TreeView?>
-<?import javafx.scene.control.TreeItem?>
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ru.gb.Chatterbox.client.ChatController">
-    <VBox visible="false" fx:id="changeNickPanel">
+<AnchorPane fx:id="anchorPane" maxHeight="539.0" maxWidth="646.0" minHeight="200.0" minWidth="200.0" prefHeight="200.0" prefWidth="200.0" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ru.gb.Chatterbox.client.ChatController">
+
+
+    <VBox fx:id="loginPanel" minHeight="200.0" minWidth="200.0" prefHeight="200.0" prefWidth="200.0">
+        <Label text="Login:">
+            <padding>
+                <Insets top="5.0" />
+            </padding>
+            <VBox.margin>
+                <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+            </VBox.margin></Label>
+        <TextField fx:id="LoginField">
+            <VBox.margin>
+                <Insets left="5.0" right="5.0" />
+            </VBox.margin></TextField>
+        <Label text="Password:">
+            <VBox.margin>
+                <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+            </VBox.margin></Label>
+        <TextField fx:id="PasswordField" onAction="#sendAuth">
+            <VBox.margin>
+                <Insets bottom="5.0" left="5.0" right="5.0" />
+            </VBox.margin>
+        </TextField>
+        <Button onAction="#sendAuth" text="Connect">
+            <VBox.margin>
+                <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+            </VBox.margin></Button>
+        <Label accessibleRole="BUTTON" alignment="TOP_RIGHT" contentDisplay="BOTTOM" onMouseClicked="#sendRegistrationWindow" prefHeight="26.0" prefWidth="150.0" text="Registration" textAlignment="RIGHT">
+            <VBox.margin>
+                <Insets bottom="5.0" left="45.0" top="25.0" />
+            </VBox.margin>
+            <opaqueInsets>
+                <Insets />
+            </opaqueInsets>
+        </Label>
+    </VBox>
+
+
+    <VBox fx:id="registrationPanel" minHeight="200.0" minWidth="200.0" prefHeight="200.0" prefWidth="200.0" visible="false">
+        <Label text="Login:">
+            <VBox.margin>
+                <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+            </VBox.margin>
+            <padding>
+                <Insets top="5.0" />
+            </padding>
+        </Label>
+        <TextField fx:id="newLoginField">
+            <VBox.margin>
+                <Insets left="5.0" right="5.0" />
+            </VBox.margin>
+        </TextField>
+        <Label text="Password:">
+            <VBox.margin>
+                <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+            </VBox.margin>
+        </Label>
+        <TextField fx:id="newPasswordField" onAction="#sendReg">
+            <VBox.margin>
+                <Insets bottom="5.0" left="5.0" right="5.0" />
+            </VBox.margin>
+        </TextField>
+        <Label text="Nickname">
+            <VBox.margin>
+                <Insets bottom="5.0" left="5.0" right="5.0" />
+            </VBox.margin>
+        </Label>
+        <TextField fx:id="newNickField" onAction="#sendReg">
+            <VBox.margin>
+                <Insets bottom="5.0" left="5.0" right="5.0" />
+            </VBox.margin>
+        </TextField>
+        <HBox prefHeight="100.0" prefWidth="200.0">
+            <Button onAction="#sendReg" text="Registration">
+                <HBox.margin>
+                    <Insets left="5.0" right="10.0" />
+                </HBox.margin>
+            </Button>
+            <Label accessibleRole="BUTTON" alignment="BOTTOM_RIGHT" contentDisplay="BOTTOM" onMouseClicked="#sendAuthorisationWindow" prefHeight="26.0" prefWidth="100.0" text="Authorization" textAlignment="RIGHT">
+                <HBox.margin>
+                    <Insets right="5.0" />
+                </HBox.margin>
+            </Label>
+        </HBox>
+    </VBox>
+
+
+
+<!--    <VBox visible="false" fx:id="changeNickPanel">
         <HBox VBox.vgrow="ALWAYS">
             <Label text="Please enter you new nickname:"/>
             <TextField fx:id="newNickField" onAction="#sendChangeNick"/>
         </HBox>
         <Button text="SEND" onAction="#sendChangeNick"/>
         <Button text="RETURN" onAction="#returnToChat"/>
-    </VBox>
-    <VBox visible="false" fx:id="changePasswordPanel">
+    </VBox>-->
+<!--    <VBox visible="false" fx:id="changePasswordPanel">
         <HBox VBox.vgrow="ALWAYS">
             <Label text="Please enter you old password:"/>
             <TextField fx:id="oldPassField" onAction="#sendChangePass"/>
@@ -36,15 +122,17 @@
         </HBox>
         <Button text="SEND" onAction="#sendChangePass"/>
         <Button text="RETURN" onAction="#returnToChat"/>
-    </VBox>
-    <VBox fx:id="loginPanel"  visible="false">
+    </VBox>-->
+<!--    <VBox fx:id="loginPanel"  visible="false">
         <Label text="Login:"/>
         <TextField fx:id="LoginField"/>
         <Label text="Password:"/>
         <TextField fx:id="PasswordField" onAction="#sendAuth"/>
         <Button text="Connect" onAction="#sendAuth"/>
-    </VBox>
-    <VBox fx:id="mainPanel" >
+    </VBox>-->
+
+
+    <VBox fx:id="mainPanel" minHeight="539.0" minWidth="646.0" prefHeight="539.0" prefWidth="646.0" visible="false">
         <MenuBar>
             <Menu text="File">
                 <MenuItem onAction="#mockAction" text="Open" />
